Remove `pysubs2.py` CLI script in favour of `python -m pysubs2`
"There should be one-- and preferably only one --obvious way to do it."

The `pysubs2.py` script was broken and I'm not completely sure how to fix
it, so I decided to remove it entirely. This does not remove any
functionality from pysubs2 itself, as the CLI continues to be availible
by executing the module via `python -m pysubs2`.

Users are encouraged to write their own wrapper scripts should they find
them preferable. This closes issue #1.