Fix for issue #12 - more liberal handling of `str` in Python 2
In Python 2, we now allow `str` (instead of `unicode`) in `SSAEvent` and
`SSAStyle` fields, as long as they do not fall outside ASCII range.
This makes working with the library more intuitive when not using
`from __future__ import unicode_literals`.

No changes are made to Python 3 behaviour, which requires everything
to be Unicode-aware `str` and not `bytes`.