Fix breakage in Python 3.11
We had a subtle bug since version 1.2.0 of the library:
`SSAStyle.primarycolor` and friends had a mutable default value,
which could lead to unwanted state sharing among user-constructed
`SSAStyle` objects. Note that it doesn't affect styles read from `.ass`
files since these will be defined explicitly and not use the shared
defaults.

Since Python 3.11 mutable default values in dataclasses is a bug,
so this broke the whole library, as reported in #56.