fix!: Only use "surrogateescape" error handling in CLI, revert to default handling in API
Changing default error handling in Python API of the library is a breaking change,
let's change it to opt-in for now.

The benefit or "surrogateescape" seems more clear in CLI, since there is currently
no feature that works with subtitle text, it only removes (some of) annoying errors
about character encoding, without breaking anything.